parsing/requirements.txt
websocket-client==1.1.0
psycopg2-binary==2.8.6
loguru